+++ SpaceShooter/SpaceShooter/sprite.h
@@ -18,7 +18,7 @@ public:
 
 	virtual ~Sprite();
 	virtual void update();
-	void draw(Graphics &graphics, int x, int y);
+	void draw(Graphics &graphics, float x, float y);
 
 	SDL_Rect sourceRect;
 	SDL_Texture* sprite;
